requirements.txt
mem0ai>=0.1.0
requests>=2.31.0
ollama>=0.2.0
openai>=1.30.0
pydantic>=2.0.0
click>=8.1.0
python-dotenv>=1.0.0
tqdm>=4.66.0
pytest>=7.4.0
pytest-asyncio>=0.23.0